// frontend/src/types/api.ts
export interface User {
  id: string;
  email: string;
  name: string;
  username: string;
  emailVerified: boolean;
  oauthProvider?: string;
  createdAt: string;
  updatedAt: string;
}